change buffer lengths of sysctl(2) to be size_t's
SCCS-vsn: 7.14

+++ sys/kern/uipc_domain.c
@@ -30,7 +30,7 @@
  * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
  * SUCH DAMAGE.
  *
- *	@(#)uipc_domain.c	7.13 (Berkeley) 03/02/93
+ *	@(#)uipc_domain.c	7.14 (Berkeley) 03/04/93
  */
 
 #include <sys/param.h>
@@ -139,9 +139,9 @@ net_sysctl(name, namelen, oldp, oldlenp, newp, newlen, p)
 	int *name;
 	u_int namelen;
 	void *oldp;
-	u_int *oldlenp;
+	size_t *oldlenp;
 	void *newp;
-	u_int newlen;
+	size_t newlen;
 	struct proc *p;
 {
 	register struct domain *dp;
